Add files via upload
This Slackbuild script unzip EORTOLOGIO GR file and place it. Down load a free icon and place desktop entry for the AppImage. You DONT have to be  root to run the script so nothing is against AppImage philosophy. BUT if  you want to load the icon on you Application menu you need as root to #update- desktop or restart DE. 
Source code included, because every release of EORTOLOGIO have different names not compatibles with SlackBuilds theory...